requirements.txt
# requirements for signald
attrs
-e git+https://github.com/mobilecoinofficial/full-service.git@d94572e82c48dc615fba046891e3f03cf8028a59#egg=mobilecoin-python-utils&subdirectory=python-utils
-e git+https://github.com/mobilecoinofficial/full-service.git@d94572e82c48dc615fba046891e3f03cf8028a59#egg=mobilecoin-cli&subdirectory=cli

base58==2.1.0
django==3.2.7
psycopg2==2.9.1
uwsgi==2.0.19.1
pyzt
googlemaps
protobuf==3.17.3

git+https://gitlab.com/stavros/pysignald.git@5c27c092fd5379f3d2af0ec21be2d915e4e3f256#egg=pysignald

django-extensions==3.1.3
aenum==3.1.0
django-phonenumber-field[phonenumberslite]==2.1.0
tenacity==8.0.1

# Make manage.py shell use an iPython session
ipython

# For tests
factory_boy==3.2.0
faker==8.11.0

# For linting
mypy==0.910
types-python-dateutil==2.8.0
types-pytz==2021.1.2
halo==0.0.31
cached-property==1.2.0
